+++ test/stripdown_render_test.rb
@@ -2,41 +2,39 @@
 require 'test_helper'
 
 class StripDownRender < Redcarpet::TestCase
-
   def setup
-    @markdown = Redcarpet::Markdown.new(Redcarpet::Render::StripDown)
+    @parser = Redcarpet::Markdown.new(Redcarpet::Render::StripDown)
   end
 
-  def test_basics
-    markdown = <<-Markdown
-# [Foo bar](https://github.com)
-Markdown
-    html = @markdown.render(markdown)
-    html_equal "Foo bar\n", html
+  def test_titles
+    markdown = "# Foo bar"
+    output   = @parser.render(markdown)
+
+    assert_equal "Foo bar\n", output
   end
 
-  def test_insert_new_lines_char
-    markdown = <<-Markdown
-# Foo bar
+  def test_code_blocks
+    markdown = "\tclass Foo\n\tend"
+    output   = @parser.render(markdown)
 
-Hello world! Please visit [this site](https://github.com/).
+    assert_equal "class Foo\nend\n", output
+  end
 
-    class Foo
-    end
+  def test_images
+    markdown = "Look at this ![picture](http://example.org/picture.png)\n" \
+               "And this: ![](http://example.org/image.jpg)"
+    expected = "Look at this picture http://example.org/picture.png\n" \
+               "And this: http://example.org/image.jpg\n"
+    output   = @parser.render(markdown)
 
-Look at this ![picture](http://example.org/picture.png)
-And this: ![](http://example.org/image.jpg)
-Markdown
-    plaintext = <<-Plaintext
-Foo bar
-Hello world! Please visit this site.
-class Foo
-end
-Look at this picture http://example.org/picture.png
-And this: http://example.org/image.jpg
-Plaintext
+    assert_equal expected, output
+  end
 
-    html = @markdown.render(markdown)
-    html_equal plaintext, html
+  def test_links
+    markdown = "Here's an [example](https://github.com)"
+    expected = "Here's an example (https://github.com)\n"
+    output   = @parser.render(markdown)
+
+    assert_equal expected, output
   end
 end