+++ ext/redcarpet/html.c
@@ -429,7 +429,10 @@ static int
 rndr_image(struct buf *ob, const struct buf *link, const struct buf *title, const struct buf *alt, void *opaque)
 {
 	struct html_renderopt *options = opaque;
-	if (!link || !link->size) return 0;
+
+
+	if (link != NULL && (options->flags & HTML_SAFELINK) != 0 && !sd_autolink_issafe(link->data, link->size))
+		return 0;
 
 	BUFPUTSL(ob, "<img src=\"");
 	escape_href(ob, link->data, link->size);