+++ lib/redcarpet/render_strip.rb
@@ -43,6 +43,18 @@ module Redcarpet
       def header(text, header_level)
         text + "\n"
       end
+
+      def table(header, body)
+        "#{header}#{body}"
+      end
+
+      def table_row(content)
+        content + "\n"
+      end
+
+      def table_cell(content, alignment)
+        content + "\t"
+      end
     end
   end
 end