+++ CHANGELOG.md
@@ -4,6 +4,11 @@
 
   *silverhammermba*
 
+* The `StripDown` render object now displays the URL of links
+  along with the text. 
+
+  *Robin Dupret*
+
 * The RedCloth API compatibility layer is now deprecated.
 
   *Robin Dupret*